+++ playwright-tests/send/actions.ts
@@ -4,9 +4,11 @@ import { DataTestIdsEnum } from '../../src/localConstants/dataTestIds.enum';
 export const completeSendForm = async (page: Page, payload) => {
   await page.getByTestId(DataTestIdsEnum.receiverInput).fill(payload.adress);
   await page.waitForTimeout(500);
-  payload.amount
-    ? await page.getByTestId(DataTestIdsEnum.amountInput).fill(payload.amount)
-    : '';
+
+  if (payload.amount) {
+    await page.getByTestId(DataTestIdsEnum.amountInput).fill(payload.amount);
+  }
+
   await page.getByTestId(DataTestIdsEnum.sendBtn).click();
 };